updateViewTitle dans utils/date.ts
